app/python/simple/pyproject.toml
[tool.poetry]
name = "rag-on-azure-hol"
version = "0.1.0"
description = ""
authors = ["kohei3110 <kk.31108424@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
openai = "^1.55.1"
requests = "^2.32.3"
azure-search-documents = "^11.5.2"
azure-identity = "^1.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
